requirements.txt
# Django
Django==4.2.7
asgiref==3.7.2
pip==22.3.1
setuptools==65.5.1
sqlparse==0.4.4
wheel==0.38.4
python-decouple==3.8
django-templated-mail==1.1.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.5
defusedxml==0.7.1
idna==3.4
pycparser==2.21
python3-openid==3.2.0
pytz==2023.3.post1
urllib3==2.1.0
django-filter==23.5
django-cors-headers==4.3.1

# DB
psycopg2-binary==2.9.9

# Auth
PyJWT==2.8.0
djangorestframework-simplejwt==5.3.0
requests==2.31.0
requests-oauthlib==1.3.1
social-auth-app-django==5.4.0
social-auth-core==4.5.0
oauthlib==3.2.2

# Restframework
djangorestframework==3.14.0
djoser==2.2.2
